requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.1
decorator==4.4.2
Django==2.2
django-grappelli==2.14.3
django-jet==1.0.8
django-phone-field==1.8.1
et-xmlfile==1.1.0
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
future==0.18.2
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
install==1.3.4
isodate==0.6.0
lxml==4.6.2
networkx==2.5
nibabel==3.2.1
nipype==1.6.0
numpy==1.19.5
openpyxl==3.0.7
packaging==20.8
pandas==1.2.1
pathlib==1.0.1
pdfrw==0.4
Pillow==8.1.0
prov==2.0.0
psycopg2==2.9.1
pydot==1.4.1
PyMuPDF==1.18.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
pyxnat==1.4
rdflib==5.0.0
requests==2.25.1
scipy==1.6.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
traits==6.2.0
urllib3==1.26.2
whitenoise==5.2.0
